Fixed all warnings related to non-matching format string parameters, when building with GCC 4.6.3 (64-bit)

+++ src/cgame/cg_main.c
@@ -90,7 +90,7 @@ Q_EXPORT intptr_t vmMain(intptr_t command, intptr_t arg0, intptr_t arg1, intptr_
 	case CG_MESSAGERECEIVED:
 		return -1;
 	default:
-		CG_Error("vmMain: unknown command %i", command);
+		CG_Error("vmMain: unknown command %li", (long)command);
 		break;
 	}
 	return -1;
